+++ lambdas/postcode-lookup/src/main/java/uk/gov/di/ipv/cri/address/api/models/Dpa.java
@@ -1,7 +1,7 @@
 package uk.gov.di.ipv.cri.address.api.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import uk.gov.di.ipv.cri.common.library.domain.CanonicalAddress;
+import uk.gov.di.ipv.cri.common.library.persistence.item.CanonicalAddress;
 
 public class Dpa {
